+++ RTSClone/RTSClone/Sprite.h
@@ -20,7 +20,7 @@ public:
 	void render(const glm::vec3& position, glm::uvec2 windowSize, float originalWidth, float spriteWidth, float height, float yOffset,
 		ShaderHandler& shaderHandler, const Camera& camera, const glm::vec3& materialColor, float opacity = 1.0f) const;
 	
-	void render(glm::ivec2 position, glm::ivec2 size, const glm::vec3& color, ShaderHandler& shaderHandler,
+	void render(glm::vec2 position, glm::vec2 size, const glm::vec3& color, ShaderHandler& shaderHandler,
 		glm::uvec2 windowSize, float opacity = 1.0f) const;
 
 private:
